FIX: batch_task_logger 함수 분리
